requirements.txt
anthropic>=0.25.0
openai>=1.30.0
pymupdf>=1.24.0
Pillow>=10.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
rich>=13.0.0
tqdm>=4.66.0
